+++ NEWS.md
@@ -1,5 +1,18 @@
-# waxer 0.0.1
+# waxer 0.9.0
 
-This is the first, early release of waxer. In this release it has:
-- Number of active editors (`wx_active_editors`)
-- Top 100 daily editors by edits (`wx_top_editors`)
+This is the first, early release of waxer. In this release it has metrics & data for:
+- **Traffic**
+  - Project view count (`wx_project_views`)
+  - Page view count (`wx_page_views`)
+  - Unique devices (`wx_unique_devices`)
+  - Top 1000 pages by views (`wx_top_viewed_pages`)
+- **Users**
+  - Number of active editors (`wx_active_editors`)
+  - Top 100 daily editors by edits (`wx_top_editors`)
+  - Newly registered users (`wx_newly_registered_users`)
+- **Content**
+  - New pages (`wx_new_pages`)
+  - Total pages (`wx_total_pages`)
+  - Project edit count (`wx_project_edits`)
+  - Page edit count (`wx_page_edits`)
+  - Top 100 edited pages by edits (`wx_top_edited_pages`)